refactor(core): 优化Java代码并重构前端组件
重构多处Java代码以提升可读性和规范性，主要包括：
1. 统一处理 longValue() 拆箱操作，直接使用 Long 类型。
2. 修复 DynamicApiCircuitBreaker 和 NoticeStreamService 中的 nullability 警告，使用 Objects.requireNonNull。
3. 更新 Java 集合的实例化方式为 Diamond Operator。
4. 调整 ExcelTool 中的字段元数据解析和值转换逻辑，清理无用代码和注释。
5. 统一日志异步执行器中的 @NonNull 注解。
6. 调整 DataScopeType、Permission、Role、RolePermission、UserRole 等实体类的 Javadoc 样式。
7. 优化 CachedBodyHttpServletRequest、DuplicateSubmitFilter、RateLimitFilter 中 Javadoc 注释和私有方法的实现。
8. 调整 DuplicateSubmitProperties 和 RateLimitProperties 中的 Javadoc 样式和属性校验逻辑。

前端组件更新：
1. 更新 Druid 监控页面，优化数据源和 SQL 列表的展示，调整国际化键名。
2. 将多处异步数据加载改为 await 语法，提升可读性。
3. 移除 dict.ts 中批量加载和按类型加载字典数据的功能，统一由 loadAll 处理。
4. 调整多个 Vue 组件的样式，清理冗余 CSS。
5. 删除 auth.ts 中冗余的 LogoutPayload 接口定义，优化接口调用顺序。
6. 简化 i18n 模块的 locale 设置逻辑。

+++ auth/src/main/java/com/example/demo/auth/store/CaptchaStore.java
@@ -80,7 +80,7 @@ public class CaptchaStore {
         }
         String key = buildCounterKey(scopeKey);
         Long count = cacheTool.increment(key);
-        long value = count == null ? 1L : count.longValue();
+        long value = count == null ? 1L : count;
         if (value == 1L) {
             cacheTool.expire(key, Duration.ofSeconds(cleanupIntervalSeconds));
         }
